Use Spring Boot standard way of configuring the datasource
First of all there's a property for that and I wouldn't manually instantiate the database If i didn't have to.

Then, all properties defined that way can be overwritten by environment properties as as shown. There's a thing call "relaxed" binding that allows to transparently use CamelCase, dashes or UPPER_AND_UNDERSCORES.

That relaxed binding is available only for ConfigurationProperties, not @Value.

Find this and more in http://springbootbuch.de ;)